Dockerfile: Set env variables `TERM` to `xterm`
Attaching to the running container `mariadb-galera` with

	sudo docker exec -ti mariadb-galera bash

and running `mysql` in it, it fails with the error below.

	root@mariadb-galera:/# mysql
	TERM environment variable not set.

So set the environment variable `TERM` in the Docker image to `xterm` so
that `mysql` can be run without any problems.

`xterm` was chosen, because it was the first one tried and it worked.
The implications are unknown.